% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryInitState.R
\name{binaryInitState}
\alias{binaryInitState}
\title{Generate all possible initial state of community}
\usage{
binaryInitState(N)
}
\arguments{
\item{N}{Total number of species in the community.}
}
\value{
A matrix of 2^NxN, with values 1 indicating the presence and 0 the
absence of the species.
}
\description{
To generate a matrix of all possible combination of species presence (1) or
absence (0) as an initial state of the community. For a community of \code{N}
species, there are \code{2^N} possibilities.
}
\examples{
# Generate all possible binary initial state of a 4 species community 
mask <- binaryInitState(4)
initAbundance <- c(1, 3, 2, 4)
# Initial abundance for all 16 scenarios
allInitAbundance <- t(t(mask) * initAbundance)

}
