% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommunitySim-package.R
\docType{package}
\name{CommunitySim-package}
\alias{CommunitySim}
\alias{CommunitySim-package}
\title{CommunitySim: Simulate an ecological community using Lotka-Volterra and its derived models}
\description{
CommunitySim simulates an ecological community, such as a microbiome using Lotka-Volterra based models. In particular, dynamics of species are simulated assuming a third species can influence the intensity of pairwise interactions, referred to as a higher-order interaction.
}
\author{
\strong{Maintainer}: Yuan Yin \email{yuan_yin@outlook.com}

}
\keyword{internal}
