% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSteadyState.R
\name{findSteadyState}
\alias{findSteadyState}
\title{Find steady state value in a vector}
\usage{
findSteadyState(vec)
}
\arguments{
\item{vec}{An atomic vector of abundance level of a species across time.}
}
\value{
A scalar where steady state has reached.
}
\description{
Given a vector as the abundance level across time, find the value in the
vector where the abundance has reached a steady state.
}
