% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCommunity.R
\name{glv}
\alias{glv}
\title{Generalized Lotka-Volterra equation}
\usage{
glv(time, x, param)
}
\arguments{
\item{time}{Time point for solver evaluation.}

\item{x}{Species abundance.}

\item{param}{Need alpha, c0 and ck. See growthFunction doc.}
}
\description{
The coupled ODE of growth rates of species in the community, with parameters
reflecting their interaction intensity. To be used as input to deSolve ODE
solver.
}
