% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCommunity.R
\name{growthFunction}
\alias{growthFunction}
\title{Simulate community}
\usage{
growthFunction(N, alpha, c0, ck, init)
}
\arguments{
\item{N}{Number of members in the community.}

\item{alpha}{1xN vector of intrinsic growth rate of each member of the
community.}

\item{c0}{NxN matrix of pairwise interaction intensity in the community.
\code{c0[i,j]} is the impact of j on i per j unit.}

\item{ck}{A list of N matrix, each with NxN entries. \code{ck[[k]][i,j]} is
the higher order impact from species j onto the pairwise interaction of
\code{i} and \code{k}, felt by \code{k}.}

\item{init}{1xN vector of initial abundance of each member in the community.}
}
\value{
A DataFrame, with row being each time point and columns species.
The abundance vs. time plot is also rendered.
}
\description{
Simulate the dynamics of the abundance of the species in the community with
parameters reflecting the interaction intensity among members of the
community.
}
\examples{
# Simulate parameters for a 4 species community
param <- simulateParam(4)
abundance <- growthFunction(param$N, param$alpha, param$c0, param$ck, 
param$init)

}
