% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateParam.R
\name{simulateParam}
\alias{simulateParam}
\title{Simulate parameters for community simulation}
\usage{
simulateParam(N)
}
\arguments{
\item{N}{Number of members in the community.}
}
\value{
A list of parameters for community simulation. \code{N} is the input.
\code{alpha} is the intrinsic growth rate. \code{c0[i,j]} is the pairwise
interaction intensity from species \code{j} to species \code{i}. \code{ck} is
a list of length \code{N}, each matrix \code{k} being the higher order
impact from species j onto the pairwise interaction of \code{i} and \code{k},
felt by \code{k}.
}
\description{
Simulate intrinsic growth rate of species in the community, and their
interaction intensity. Interaction intensity are mostly negative to reflect
the competitive nature in the community, and values are drawn from a uniform
distribution.
}
\examples{
re <- simulateParam(5)
alpha <- re$alpha

}
